Warn about shallow pointers in user-defined types
While we already emit a warning for pointers to any foreign types, the
final hurdle to guarantee full EDL serialization is the use of any
pointer in a user-defined type that does not have a size or count
attribute specified, meaning the edger8r will copy the pointer by value
instead of performing a deep-copy. So instead we emit a warning for
these fields.

Signed-off-by: Andrew Schwartzmeyer <andrew@schwartzmeyer.com>